requirements.txt
click==8.3.0
pydantic==2.11.8
PyYAML==6.0.3